refactor: test refactor
