+++ src/main/kotlin/ru/dnlkk/tagger/TaggerApplication.kt
@@ -3,26 +3,25 @@ package ru.dnlkk.tagger
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import org.springframework.context.annotation.Bean
+import ru.dnlkk.tagger.configuration.TaggerProperties
 import ru.dnlkk.tagger.infrastructure.front_controller.TaggerFrontController
 
 @SpringBootApplication
 class TaggerApplication {
+
     @Bean
-    fun tagger(): Tagger {
-        val tagger = Tagger(taggerProperties(), frontController())
-        tagger.startPolling()
-        return tagger
+    fun tagger(
+        properties: TaggerProperties,
+        taggerFrontController: TaggerFrontController
+    ) = Tagger(properties, taggerFrontController).apply {
+        startPolling()
     }
 
     @Bean
-    fun taggerProperties(): TaggerProperties {
-        return TaggerProperties()
-    }
+    fun taggerProperties() = TaggerProperties()
 
     @Bean
-    fun frontController(): TaggerFrontController {
-        return TaggerFrontController()
-    }
+    fun frontController() = TaggerFrontController()
 }
 
 fun main() {
